FM-173: Arbitraries for IPC message
