+++ fendermint/vm/genesis/Cargo.toml
@@ -38,4 +38,11 @@ fvm_ipld_encoding = { workspace = true }
 
 [features]
 default = []
-arb = ["arbitrary", "quickcheck", "fvm_shared/arb", "fendermint_testing/arb", "rand", "cid"]
+arb = [
+  "arbitrary",
+  "quickcheck",
+  "fvm_shared/arb",
+  "fendermint_testing/arb",
+  "rand",
+  "cid",
+]
