+++ fendermint/vm/genesis/src/arb.rs
@@ -4,7 +4,7 @@ use crate::{
     ipc, Account, Actor, ActorMeta, Genesis, Multisig, Power, SignerAddr, Validator, ValidatorKey,
 };
 use cid::multihash::MultihashDigest;
-use fendermint_testing::arb::ArbTokenAmount;
+use fendermint_testing::arb::{ArbSubnetID, ArbTokenAmount};
 use fendermint_vm_core::Timestamp;
 use fvm_shared::{
     address::Address,
@@ -12,7 +12,6 @@ use fvm_shared::{
     econ::TokenAmount,
     version::NetworkVersion,
 };
-use ipc_sdk::subnet_id::SubnetID;
 use quickcheck::{Arbitrary, Gen};
 use rand::{rngs::StdRng, SeedableRng};
 
@@ -98,29 +97,6 @@ impl Arbitrary for Genesis {
                 None
             },
         }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct ArbSubnetID(SubnetID);
-
-impl Arbitrary for ArbSubnetID {
-    fn arbitrary(g: &mut quickcheck::Gen) -> Self {
-        let child_count = usize::arbitrary(g) % 4;
-
-        let children = (0..child_count)
-            .map(|_| {
-                if bool::arbitrary(g) {
-                    Address::new_id(u64::arbitrary(g))
-                } else {
-                    // Only expectign EAM managed delegated addresses.
-                    let subaddr: [u8; 20] = std::array::from_fn(|_| Arbitrary::arbitrary(g));
-                    Address::new_delegated(10, &subaddr).unwrap()
-                }
-            })
-            .collect::<Vec<_>>();
-
-        Self(SubnetID::new(u64::arbitrary(g), children))
     }
 }
