+++ fendermint/vm/message/golden/chain/ipc_bottom_up.cbor
@@ -1,0 +1,1 @@
+a163497063a168426f74746f6d5570a2676d657373616765a3676d657373616765a26a636865636b706f696e74a4697375626e65745f6964821bffffffffffffffff8356040a150a15a7f73119c5fff59f4536fdff1e6628f8924b00bd8ac9ece6a196ae8c014b0088ee8bacc9ab83cbea01666865696768741a61f7b23f756e6578745f76616c696461746f725f7365745f69641bab4526659dc9c1d272626f74746f6d5f75705f6d65737361676573d82a5823001220d516016a37df05d05b0bc1d0859e6ccaef3a39f98cf83122def22df12a517fa06a7369676e61747572657383a26976616c696461746f725502add901ff4cbd310000b279519409aede7b206bd2697369676e61747572654101a26976616c696461746f724a00c790fb9dddacdfe92f697369676e61747572654102a26976616c696461746f725004d8df80b1a5fafdf5fa0100464f5614697369676e61747572654701bb3bb728ed506772656c617965725831032e4b673d0e6f8bcfb6015952eabe29e23a05bbba00f63e5ee03f91ffa5484868b794740fd2e7d3a21a000197ffd0b6006873657175656e63651b9345babad4844692697369676e61747572654301ffa3
