+++ fendermint/vm/message/golden/chain/ipc_bottom_up.txt
@@ -1,0 +1,1 @@
+Ipc(BottomUp(SignedRelayedMessage { message: RelayedMessage { message: SignedBottomUpCheckpoint { checkpoint: BottomUpCheckpoint { subnet_id: SubnetID { root: 18446744073709551615, children: [Address { payload: Delegated(DelegatedAddress { namespace: 10, length: 20, buffer: [21, 10, 21, 167, 247, 49, 25, 197, 255, 245, 159, 69, 54, 253, 255, 30, 102, 40, 248, 146, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0] }) }, Address { payload: ID(10114056781670008125) }, Address { payload: ID(16903712942678275848) }] }, height: 1643622975, next_validator_set_id: 12341312571759509970, bottom_up_messages: Cid(QmcgT1b8CJhRtmqFsnX2QVUmiMYLFwcViQGVbAi9UWfyRq) }, signatures: [ValidatorSignature { validator: Address { payload: Actor([173, 217, 1, 255, 76, 189, 49, 0, 0, 178, 121, 81, 148, 9, 174, 222, 123, 32, 107, 210]) }, signature: Signature { sig_type: Secp256k1, bytes: [] } }, ValidatorSignature { validator: Address { payload: ID(3446236016166225991) }, signature: Signature { sig_type: BLS, bytes: [] } }, ValidatorSignature { validator: Address { payload: Delegated(DelegatedAddress { namespace: 18080817611769130968, length: 5, buffer: [0, 70, 79, 86, 20, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0] }) }, signature: Signature { sig_type: Secp256k1, bytes: [187, 59, 183, 40, 237, 80] } }] }, relayer: Address { payload: BLS([46, 75, 103, 61, 14, 111, 139, 207, 182, 1, 89, 82, 234, 190, 41, 226, 58, 5, 187, 186, 0, 246, 62, 94, 224, 63, 145, 255, 165, 72, 72, 104, 183, 148, 116, 15, 210, 231, 211, 162, 26, 0, 1, 151, 255, 208, 182, 0]) }, sequence: 10612093408560563858 }, signature: Signature { sig_type: Secp256k1, bytes: [255, 163] } }))
