+++ fendermint/vm/message/golden/chain/ipc_top_down.txt
@@ -1,0 +1,1 @@
+Ipc(TopDown)
